Fix some typos.
git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@45336 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/network/src/java/org/apache/geronimo/network/protocol/util/IntegerKeyUpPacket.java
@@ -21,13 +21,13 @@ import org.apache.geronimo.network.protocol.UpPacket;
 
 
 /**
- * @version $Revision: 1.2 $ $Date: 2004/03/10 09:59:15 $
+ * @version $Revision: 1.3 $ $Date: 2004/04/20 12:16:54 $
  */
 public abstract class IntegerKeyUpPacket extends UpPacket {
 
     final int key;
 
-    public IntegerKeyUpPacket(short key) {
+    public IntegerKeyUpPacket(int key) {
         this.key = key;
     }
 }
